[PC-498] Chore: Login, VerifingContact View 폴더링
